perf: 优化最后更新
