+++ apps/common/signal_handlers.py
@@ -1,15 +1,16 @@
 # -*- coding: utf-8 -*-
 #
-import re
-import os
 import logging
+import re
 from collections import defaultdict
+
 from django.conf import settings
 from django.core.signals import request_finished
 from django.db import connection
+from django.db.models.signals import pre_save
+from django.dispatch import receiver
 
 from jumpserver.utils import get_current_request
-
 from .local import thread_local
 
 pattern = re.compile(r'FROM `(\w+)`')
@@ -81,6 +82,36 @@ def on_request_finished_logging_db_query(sender, **kwargs):
 
 def on_request_finished_release_local(sender, **kwargs):
     thread_local.__release_local__()
+
+
+def _get_request_user_name():
+    user_name = 'System'
+    current_request = get_current_request()
+    if current_request and current_request.user.is_authenticated:
+        user_name = current_request.user.name
+        if isinstance(user_name, str):
+            user_name = user_name[:30]
+    return user_name
+
+
+@receiver(pre_save)
+def on_create_set_created_by(sender, instance=None, **kwargs):
+    if getattr(instance, '_ignore_auto_created_by', False):
+        return
+    if not hasattr(instance, 'created_by') or instance.created_by:
+        return
+    user_name = _get_request_user_name()
+    instance.created_by = user_name
+
+
+@receiver(pre_save)
+def on_update_set_updated_by(sender, instance=None, created=False, **kwargs):
+    if getattr(instance, '_ignore_auto_updated_by', False):
+        return
+    if not hasattr(instance, 'updated_by'):
+        return
+    user_name = _get_request_user_name()
+    instance.updated_by = user_name
 
 
 if settings.DEBUG_DEV:
